Factor PropagatedState and settings into their own file
